CreateArcGISFeatureLayer.py - still not working, but now trying to use QgsProcessingOutputVectorLayer as the output type.
feedback displays that the layer is valid, and does have a feature count greater than zero. However, when this processor is used in anger, the downstream processor says that the layer (i.e. the output) can not be found.

ERROR:
Could not load source layer for INPUT: RT_a269883f_03c1_4149_9190_89cacf160f28 not found